The list of users can now be saved to and loaded from a text file.
We store the users in an ArrayList<User>. When the users are loaded from the file, the ArrayList is populated. When a new user data is inserted, it is added to the ArrayList. When we wan't to save the list of users, the file is overwritten with all the data in the ArrayList.